+++ src/constants/strings.ts
@@ -2,6 +2,7 @@ export type PagePathname
     = '/'
     | '/#description'
     | '/recommendations'
+    | '/search'
     | `/manage${string}`
     | `/users${string}`
     | `/dishes${string}`;